CPR-615 Add extra logic
