--- src/main/kotlin/uk/gov/justice/digital/hmpps/personrecord/service/message/PersonMatchReclusterService.kt
@@ -1,16 +1,0 @@
-package uk.gov.justice.digital.hmpps.personrecord.service.message
-
-import org.springframework.stereotype.Component
-import uk.gov.justice.digital.hmpps.personrecord.jpa.entity.PersonKeyEntity
-import uk.gov.justice.digital.hmpps.personrecord.service.search.PersonMatchService
-
-@Component
-class PersonMatchReclusterService(
-  private val personMatchService: PersonMatchService
-) {
-
-  fun recluster(personKeyEntity: PersonKeyEntity) {
-
-  }
-
-}
